feature/PRK-2: add API create book

+++ src/main/java/com/example/parkee_book_library/repository/BorrowerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BorrowerRepository extends JpaRepository<Borrower, Integer> {
+public interface BorrowerRepository extends JpaRepository<Borrower, Long> {
 }
